feat: implement simple real-time message sending
The code introduced in this commit is to test the
sending and receiving of messages in real-time
using socket.io. This was written while working
on my personal Desktop, but eventually I need to
update this code so that it works will on my
Virtual Machine that uses Nginx as the reverse
proxy.

Next I'll need to introduce a Database to start
storing messages and will need to implement
other features of socket.io such as rooms and
joins.